Update Task and File management system.py
Task management is for adding task with an option to mark the task as important with the date and time, display the task and mark the task as completed when done or can delete the task.

File management is like keeping track of all the files on your computer, user can create file with optional password for important file and user can read the file only if  entered correct password and username. User can add content to file and delete the file by entering the correct password.